Reorganized scalalib, more files, some docs, slighly refactored SbtModule
